Move frontend to /client
